+++ spec/schleuder/unit/keyword_handlers_runner_spec.rb
@@ -34,21 +34,10 @@ describe 'KeywordHandlersRunner' do
     expect(output).to eql(['Your message contained an incorrect "X-LIST-NAME" keyword. The keyword argument must match the email address of this list.'])
   end
 
-  it 'requires X-STOP' do
-    list = create(:list)
-    mail = Mail.new
-    mail.body = "x-list-keys\nx-list-name: #{list.email}"
-    mail.to_s
-
-    output = KeywordHandlersRunner.run(mail: mail, list: list, type: :request)
-
-    expect(output).to eql(["Your message lacked the keyword 'X-STOP'. If you use keywords in a message, you must indicate with 'X-STOP' where to stop looking for further keywords."])
-  end
-
   it 'rejects unknown keywords' do
     list = create(:list)
     mail = Mail.new
-    mail.body = "x-list-subscriptions\nx-blabla\nx-list-name: #{list.email}\nx-stop"
+    mail.body = "x-list-subscriptions\nx-blabla\nx-list-name: #{list.email}"
     mail.to_s
 
     output = KeywordHandlersRunner.run(mail: mail, list: list, type: :request)
@@ -67,7 +56,7 @@ describe 'KeywordHandlersRunner' do
     mail.list.subscribe('subscription@example.net', 'C4D60F8833789C7CAA44496FD3FFA6613AB10ECE', false)
     mail.list.import_key(File.read('spec/fixtures/example_key.txt'))
     mail.instance_variable_set('@signing_key', mail.list.key('C4D60F8833789C7CAA44496FD3FFA6613AB10ECE'))
-    mail.body = "x-custom-keyword\nx-list-name: #{mail.list.email}\nx-stop"
+    mail.body = "x-custom-keyword\nx-list-name: #{mail.list.email}"
     mail.to_s
 
     output = KeywordHandlersRunner.run(mail: mail, list: mail.list, type: :request)
@@ -82,7 +71,7 @@ describe 'KeywordHandlersRunner' do
     mail.list.subscribe('schleuder@example.org', '59C71FB38AEE22E091C78259D06350440F759BD3', true)
     mail.list.import_key(File.read('spec/fixtures/example_key.txt'))
     mail.instance_variable_set('@signing_key', mail.list.key('59C71FB38AEE22E091C78259D06350440F759BD3'))
-    mail.body = "x-list-subscriptions\nx-list-name: #{mail.list.email}\nx-stop"
+    mail.body = "x-list-subscriptions\nx-list-name: #{mail.list.email}"
     mail.to_s
 
     output = KeywordHandlersRunner.run(mail: mail, list: mail.list, type: :request)
@@ -106,7 +95,7 @@ describe 'KeywordHandlersRunner' do
     mail.list.subscribe('subscription@example.org', '59C71FB38AEE22E091C78259D06350440F759BD3', false)
     mail.list.import_key(File.read('spec/fixtures/example_key.txt'))
     mail.instance_variable_set('@signing_key', mail.list.key('59C71FB38AEE22E091C78259D06350440F759BD3'))
-    mail.body = "x-list-subscriptions\nx-list-name: #{mail.list.email}\nx-stop"
+    mail.body = "x-list-subscriptions\nx-list-name: #{mail.list.email}"
     mail.to_s
 
     output = KeywordHandlersRunner.run(mail: mail, list: mail.list, type: :request)